Fix T48456, where retina displays image paint strokes are painted double the UnifiedBrushSize width (i.e. size * U.pixelsize). As a result there was no way to paint 1 pixel width strokes in the image editor.
It looks like the problem was that the sculpt/paint brush cursor effective area was "adjusted" for retina by inflating the effective size of the brush by U.pixelsize. However, this meant the smallest brush you could ever paint with was 1 * U.pixelsize. My patch changes the way the brush is adjusted for retina, by making it inflate the drawn brush cursor to account for U.pixelsize. This way, the unified brush size is used directly for painting, and 1 pixel strokes come out 1 pixel large on retina machines.