Add a blender "Unsaved changes exist?" confirm menu when some changes are unsaved during open, open recent, and file-drop operations.
This dialog only occurs on the open_mainfile operator INVOKE, since it's responding to user-action. open-recent and file-drop were previously treated as EXEC, now treated as INVOKE, since they are initiated by user-action. INVOKE now conditional triggers a file choose based on whether a "filepath" was supplied to the operator.